Deprecate old configurations
At current commit, the code is not testable.

In order to showcase the deprecated options, this commit only
includes the logic for the deprecated options and does not
include the logic for migrating the configuration yet.

In the next commit, the code will be returned to a testable state.

---

A preview for current configuration:

config :ex_money,
  default_cldr_backend: Money.Cldr,
  json_library: Jason,
  exchange_rates: [
    api_module: Money.ExchangeRates.OpenExchangeRates,
    callback_module: Money.ExchangeRates.Callback,
    cache_module: Money.ExchangeRates.Cache.Dets,
    auto_start: false,
    retrieve_every: 300000,
    log_level_success: :info,
    log_level_failure: :warning,
    log_level_info: :info,
    verify_peer: true
  ],
  open_exchange_rates_app_id: {:system, "OPEN_EXCHANGE_RATES_APP_ID"}
]